+++ src/Services/Products.php
@@ -5,8 +5,8 @@ use GuzzleHttp\Psr7\Response;
 use MPAPI\Endpoints\ProductsEndpoints;
 use MPAPI\Endpoints\VariantsEndpoints;
 use MPAPI\Endpoints\Products\SupplyDelayEndpoints;
-use MPAPI\Entity\BasicProductIterator;
-use MPAPI\Entity\Product;
+use MPAPI\Entity\Products\BasicProductIterator;
+use MPAPI\Entity\Products\Product;
 use MPAPI\Entity\AbstractEntity;
 use MPAPI\Exceptions\ApplicationException;
 use MPAPI\Exceptions\EndpointNotfoundException;
@@ -26,39 +26,6 @@ class Products extends AbstractService
 	 * @var string
 	 */
 	const ENDPOINT_NAME_PATTERN = 'MPAPI\Endpoints\%sEndpoints';
-
-	/**
-	 *
-	 * @var string
-	 */
-	const ARGUMENT_FILTER = 'filter';
-
-	/**
-	 *
-	 * @var string
-	 */
-	const FILTER_TYPE_IDS = 'ids';
-
-	/**
-	 *
-	 * @var string
-	 */
-	const FILTER_TYPE_BASIC = 'basic';
-
-	/**
-	 *
-	 * @var array
-	 */
-	private $filterType = [
-		self::FILTER_TYPE_IDS,
-		self::FILTER_TYPE_BASIC
-	];
-
-	/**
-	 *
-	 * @var Client
-	 */
-	private $client;
 
 	/**
 	 *
@@ -275,34 +242,6 @@ class Products extends AbstractService
 	public function supplyDelay($productId)
 	{
 		return new SupplyDelayEndpoints($this->client, $productId);
-	}
-
-	/**
-	 * Set filter to modify response data structure
-	 *
-	 * @param string $filterType
-	 * @return Products
-	 */
-	public function setFilter($filterType)
-	{
-		if (in_array($filterType, $this->filterType)) {
-			$this->client->setArguments(self::ARGUMENT_FILTER, $filterType);
-		}
-		return $this;
-	}
-
-	/**
-	 * Get filter
-	 *
-	 * @return string
-	 */
-	public function getFilter()
-	{
-		$retval = $this->client->getArguments(self::ARGUMENT_FILTER);
-		if (empty($retval)) {
-			$retval = self::FILTER_TYPE_IDS;
-		}
-		return $retval;
 	}
 
 	/**